git-svn-id: svn+ssh://rfs.csl.sri.com/svn/internal/ashish-tools@61 598a3e1c-6b6c-e011-a152-003048836090

+++ src/polyrep.py
@@ -16,7 +16,8 @@ def nameExpr2poly(node):
 
 def numeral2poly(node):
     numstr = valueOf(node)
-    num  = int(numstr)
+    # num  = int(numstr)
+    num  = float(numstr)
     if num == 0:
         return [ ]
     return [ [ num, dict() ] ]
